Improved steering

+++ editor/src/com/starsailor/editor/util/FormUtil.java
@@ -326,7 +326,7 @@ public class FormUtil {
               else if(field.getType().equals(float.class)) {
                 value = Float.parseFloat(newValue);
               }
-              else if(field.getType().equals(float.class)) {
+              else if(field.getType().equals(int.class)) {
                 value = Integer.parseInt(newValue);
               }
               field.set(data, value);
